requirements.txt
Django==2.2.2
django-cors-headers==3.0.2
django-filter==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==1.7.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
